Move ArticleIterator to its own module
