+++ update-bot/wlmbots/update_bot.py
@@ -13,7 +13,8 @@ import mwparserfromhell
 
 from wlmbots.lib.commonscat_mapper import CommonscatMapper
 from wlmbots.lib.template_replacer import TemplateReplacer
-from wlmbots.lib.pagelist import Pagelist, ArticleIterator, ArticleIteratorArgumentParser
+from wlmbots.lib.pagelist import Pagelist
+from wlmbots.lib.article_iterator import ArticleIterator, ArticleIteratorArgumentParser
 
 
 WLM_PLACEHOLDER = '<-- link to commons placeholder "#commonscat#" -->'  # TODO proper placeholder
